requirements2.txt
alabaster==0.7.12
Babel==2.6.0
bdp==0.2.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==0.12.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.2
idna==2.8
imagesize==1.1.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
numpy==1.16.0
opencv-python==4.1.0.25
packaging==19.0
pandas==0.24.0
pexpect==4.7.0
Pillow==5.3.0
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
seaborn==0.7.1
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinx-argparse==0.2.5
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.17
urllib3==1.24.3
webcolors==1.8.1
Werkzeug==0.14.1
